BackendContainer/src/security/auth.py
from datetime import datetime, timedelta, timezone
from typing import Any, Dict, Optional, Tuple

from jose import JWTError, jwt
from passlib.context import CryptContext

from src.config import get_settings

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")


# PUBLIC_INTERFACE
def hash_password(plain_password: str) -> str:
    """Hash a plain password using bcrypt."""
    return pwd_context.hash(plain_password)


# PUBLIC_INTERFACE
def verify_password(plain_password: str, hashed_password: str) -> bool:
    """Verify a plain password against a bcrypt hash."""
    return pwd_context.verify(plain_password, hashed_password)


# PUBLIC_INTERFACE
def create_access_token(
    subject: str,
    tenant_id: str,
    additional_claims: Optional[Dict[str, Any]] = None,
    expires_in: Optional[int] = None,
) -> str:
    """Create a signed JWT access token including tenant and subject (user id)."""
    settings = get_settings()
    expire_seconds = expires_in if expires_in is not None else settings.JWT_EXPIRES_IN
    now = datetime.now(timezone.utc)
    payload: Dict[str, Any] = {
        "sub": subject,
        "tenant_id": tenant_id,
        "iat": int(now.timestamp()),
        "exp": int((now + timedelta(seconds=expire_seconds)).timestamp()),
        "nbf": int(now.timestamp()),
    }
    if additional_claims:
        payload.update(additional_claims)
    token = jwt.encode(payload, settings.JWT_SECRET, algorithm=settings.JWT_ALGORITHM)
    return token


# PUBLIC_INTERFACE
def decode_token(token: str) -> Tuple[bool, Optional[Dict[str, Any]], Optional[str]]:
    """Decode and validate a JWT. Returns (ok, claims, error)."""
    settings = get_settings()
    try:
        claims = jwt.decode(token, settings.JWT_SECRET, algorithms=[settings.JWT_ALGORITHM])
        return True, claims, None
    except JWTError as e:
        return False, None, str(e)
