BackendContainer/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "device-remote-management-backend"
version = "0.1.0"
description = "Backend services for Device Remote Management platform (FastAPI + async PostgreSQL)"
requires-python = ">=3.10"
authors = [{ name = "Device Remote Management Team" }]
dependencies = [
  "fastapi==0.115.5",
  "uvicorn[standard]==0.32.1",
  "pydantic==2.9.2",
  "pydantic-settings==2.6.1",
  "python-jose[cryptography]==3.3.0",
  "passlib[bcrypt]==1.7.4",
  "bcrypt==4.2.0",
  "SQLAlchemy==2.0.36",
  "asyncpg==0.30.0",
  "alembic==1.14.0",
  "python-multipart==0.0.17",
  "orjson==3.10.12",
  "psycopg[binary]==3.2.3"
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
